Cleanup activation to remove dead code and improve readability
